+++ src/main/java/com/wellseecoding/server/infra/jwt/JweMapperCreateException.java
@@ -1,4 +1,4 @@
-package com.wellseecoding.server.security.jwt;
+package com.wellseecoding.server.infra.jwt;
 
 public class JweMapperCreateException extends RuntimeException {
     public JweMapperCreateException(Throwable cause) {
